Package refactoring
Minor package refactoring for some item classes.
