+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/block/SpotlightBlock.java
@@ -74,7 +74,7 @@ public class SpotlightBlock extends HorizontalDirectionalBlock implements Simple
 	@Override
 	public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState,
 	                              LevelAccessor levelAccessor, BlockPos currentPos, BlockPos facingPos) {
-		
+
 		return facing.getOpposite() == stateIn.getValue(FACING) && !stateIn.canSurvive(levelAccessor, currentPos)
 				? Blocks.AIR.defaultBlockState() : stateIn;
 	}
