+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/entity/ai/goal/RangedGunAttackGoal.java
@@ -5,7 +5,7 @@ import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.entity.monster.RangedAttackMob;
 import net.minecraft.world.entity.projectile.ProjectileUtil;
-import tech.anonymoushacker1279.immersiveweapons.item.projectile.gun.AbstractGunItem;
+import tech.anonymoushacker1279.immersiveweapons.item.gun.AbstractGunItem;
 
 import java.util.EnumSet;
