+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/item/gun/FlareGunItem.java
@@ -1,4 +1,4 @@
-package tech.anonymoushacker1279.immersiveweapons.item.projectile.gun;
+package tech.anonymoushacker1279.immersiveweapons.item.gun;
 
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.item.Item;
