--- src/main/java/tech/anonymoushacker1279/immersiveweapons/item/projectile/throwable/package-info.java
@@ -1,9 +1,0 @@
-@ParametersAreNonnullByDefault
-@MethodsReturnNonnullByDefault
-@FieldsAreNonnullByDefault
-package tech.anonymoushacker1279.immersiveweapons.item.projectile.throwable;
-
-import net.minecraft.FieldsAreNonnullByDefault;
-import net.minecraft.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
